Fix usdentrytest
In Qt6 QVector::size() returns a qsizetype, on my system it's a qint64,
in UDSEntry load()/save() we beging de/serializing the data with "size"
which is a quint32, so use that explicitly here to make the test pass.

+++ autotests/udsentrytest.cpp
@@ -154,7 +154,7 @@ void UDSEntryTest::testSaveLoad()
     {
         QDataStream stream(&data, QIODevice::WriteOnly);
         for (const QVector<UDSTestField> &testCase : testCases) {
-            stream << testCase.count();
+            stream << static_cast<quint32>(testCase.count());
 
             for (const UDSTestField &field : testCase) {
                 uint uds = field.m_uds;
